+++ lib/load.php
@@ -169,6 +169,7 @@ require __DIR__ . '/compat/wordpress-6.6/post.php';
 
 // WordPress 6.7 compat.
 require __DIR__ . '/compat/wordpress-6.7/blocks.php';
+require __DIR__ . '/compat/wordpress-6.7/block-bindings.php';
 
 // Experimental features.
 require __DIR__ . '/experimental/block-editor-settings-mobile.php';